refactor: extract common chunk
