+++ package.json
@@ -88,7 +88,7 @@
     "prettier": "^3.5.3"
   },
   "typeCoverage": {
-    "atLeast": 92,
+    "atLeast": 92.1,
     "cache": true,
     "detail": true,
     "ignoreAsAssertion": true,
