+++ src/index.ts
@@ -1,6 +1,7 @@
 import { getMigrate } from './helpers.js'
 
 export * from './ajv.js'
+export * from './common.js'
 export * from './helpers.js'
 export type * from './types.js'
